+++ needTodo.md
@@ -4,3 +4,6 @@
 2. Banner for displaying the how many goats being deployed and remaining as well.
 3. websocket integration for two players
 4. manage file system for the project. [everything is in the same place right now]
+5. you can select the picture, if its cute or scary - must be funny either way ## fiance's request
+6. selectable landscape
+7.
